+++ server/src/models/query_options.rs
@@ -2,14 +2,14 @@ use serde::Deserialize;
 
 #[derive(Debug, Clone, Deserialize)]
 pub struct QueryOptions {
-    pub offset: Option<u32>,
+    pub offset: Option<i64>,
     pub search: Option<String>,
-    pub limit: Option<u32>,
+    pub limit: Option<i64>,
     pub order: Option<String>,
 }
 
 impl QueryOptions {
-    pub fn get_offset(&self) -> u32 {
+    pub fn get_offset(&self) -> i64 {
         self.offset.unwrap_or(0)
     }
